Improved: changed actions into methods for fetching reasons and updating inventory (#85zrx66zh)
