+++ src/locales/en.json
@@ -1,7 +1,6 @@
 {
   "App": "App",
   "Are you sure you want to change the time zone to?": "Are you sure you want to change the time zone to {timeZoneId}?",
-  "Are you sure you want to update the inventory variance?": "Are you sure you want to update the inventory variance?",
   "Camera permission denied.": "Camera permission denied.",
   "Cancel": "Cancel",
   "Change": "Change",
@@ -46,6 +45,7 @@
   "eCom Store": "eCom Store",
   "store name":"store name",
   "Results": "Results",
+  "The inventory will be immediately updated and cannot be undone. Are you sure you want to update the inventory variance?": "The inventory will be immediately updated and cannot be undone. Are you sure you want to update the inventory variance?",
   "The products in the upload list will be removed.": "The products in the upload list will be removed.",
   "The timezone you select is used to ensure automations you schedule are always accurate to the time you select.": "The timezone you select is used to ensure automations you schedule are always accurate to the time you select.",
   "This is the name of the OMS you are connected to right now. Make sure that you are connected to the right instance before proceeding.": "This is the name of the OMS you are connected to right now. Make sure that you are connected to the right instance before proceeding.",
