+++ src/store/RootState.ts
@@ -1,5 +1,4 @@
 export default interface RootState {
     user: any;
     product: any;
-    util: any;
 }
