+++ src/store/index.ts
@@ -6,7 +6,6 @@ import RootState from './RootState'
 import createPersistedState from "vuex-persistedstate";
 import userModule from './modules/user';
 import productModule from "./modules/product"
-import utilModule from "./modules/util"
 
 
 // TODO check how to register it from the components only
@@ -32,8 +31,7 @@ const store = createStore<RootState>({
     plugins: [ persistState ],
     modules: { 
         'user': userModule,
-        'product': productModule,
-        'util': utilModule,
+        'product': productModule
     },
 })
