--- src/store/modules/util/UtilState.ts
@@ -1,3 +1,0 @@
-export default interface UtilState {
-  varianceReasons: [];
-}
