--- src/store/modules/util/index.ts
@@ -1,18 +1,0 @@
-import actions from './actions'
-import getters from './getters'
-import mutations from './mutations'
-import { Module } from 'vuex'
-import UserState from './UtilState'
-import RootState from '@/store/RootState'
-
-const userModule: Module<UserState, RootState> = {
-    namespaced: true,
-    state: {
-      varianceReasons: []
-    },
-    getters,
-    actions,
-    mutations,
-}
-
-export default userModule;
